requirements.txt
langchain
openai
langchain-openai
langchain_community
python-dotenv